requirements.txt
psycopg2-binary
openpyxl
python-dotenv
pandas